+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <application
         android:name=".HabitApplication"
         android:allowBackup="true"
@@ -19,6 +17,7 @@
         <activity
             android:name=".FirstActivity"
             android:exported="true"
+            android:label="@string/app_name"
             android:theme="@style/Theme.Doubletapp_hw">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
