+++ app/src/main/java/com/doubletapp_hw/FirstActivity.kt
@@ -1,91 +1,172 @@
 package com.doubletapp_hw
 
 import android.app.Application
+import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.annotation.RequiresApi
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material3.DrawerValue
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ModalDrawerSheet
+import androidx.compose.material3.ModalNavigationDrawer
+import androidx.compose.material3.NavigationDrawerItem
+import androidx.compose.material3.NavigationDrawerItemDefaults
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.compositionLocalOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.core.view.WindowCompat
-import androidx.navigation.NavController
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
-import com.doubletapp_hw.apiUsage.HabitApiService
-import com.doubletapp_hw.apiUsage.RetrofitClient
 import com.doubletapp_hw.db.AppDatabase
+import com.doubletapp_hw.screens.HabitEditScreen
+import com.doubletapp_hw.screens.HabitsPagerScreen
+import com.doubletapp_hw.screens.InfoScreen
 import com.doubletapp_hw.screens.Routes
-import com.doubletapp_hw.screens.edit.HabitEditScreen
-import com.doubletapp_hw.screens.home.HomeScreen
-import com.doubletapp_hw.screens.info.InfoScreen
 import com.doubletapp_hw.ui.theme.Dobletapp_hwTheme
+import kotlinx.coroutines.launch
 
 class HabitApplication : Application() {
-    lateinit var database: AppDatabase
-        private set
-    lateinit var habitRepository: HabitRepository
-        private set
-    lateinit var api: HabitApiService
-        private set
-
-    val localNavController = compositionLocalOf<NavController> { error("No NavController found!") }
-
-    override fun onCreate() {
-        super.onCreate()
-        database = AppDatabase.getDatabase(this)
-        api = RetrofitClient.habitApi
-        habitRepository = HabitRepository(
-            database.habitDao(),
-            api = api,
-            token = BuildConfig.API_TOKEN,
-            context = applicationContext
-        )
-    }
+    val database: AppDatabase by lazy { AppDatabase.getDatabase(this) }
+    val habitRepository: HabitRepository by lazy { HabitRepository(database.habitDao()) }
 }
 
 class FirstActivity : ComponentActivity() {
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
         setContent {
             Dobletapp_hwTheme {
                 Surface(
-                    modifier = Modifier.fillMaxSize()
+                    modifier = Modifier.fillMaxSize(),
                 ) {
-                    AppNav()
+                    AppNavigation()
                 }
             }
         }
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AppNav() {
+fun AppNavigation() {
     val navController = rememberNavController()
-    val app = LocalContext.current.applicationContext as HabitApplication
+    val scope = rememberCoroutineScope()
+    val drawerState = rememberDrawerState(DrawerValue.Closed)
 
-    CompositionLocalProvider(app.localNavController provides navController) {
-        NavHost(
-            navController = navController,
-            startDestination = Routes.Home
-        ) {
-            composable<Routes.Home> {
-                HomeScreen()
+    val items = listOf(Routes.Home, Routes.Info)
+    var selectedItem by remember { mutableStateOf(items[0]) }
+    var showBottomSheet by remember { mutableStateOf(false) }
+
+    ModalNavigationDrawer(
+        drawerState = drawerState,
+        drawerContent = {
+            ModalDrawerSheet {
+                Spacer(Modifier.height(12.dp))
+                items.forEach { item ->
+                    NavigationDrawerItem(
+                        icon = { Icon(item.icon, contentDescription = null) },
+                        label = { Text(stringResource(item.title)) },
+                        selected = item == selectedItem,
+                        onClick = {
+                            scope.launch { drawerState.close() }
+                            selectedItem = item
+                            navController.navigate(item) {
+                                popUpTo(0)
+                            }
+                        },
+                        modifier = Modifier.padding(NavigationDrawerItemDefaults.ItemPadding)
+                    )
+                }
             }
-            composable<Routes.Info> {
-                InfoScreen()
-            }
-            composable<Routes.HabitEdit> { backStackEntry ->
-                val edit: Routes.HabitEdit = backStackEntry.toRoute()
-                HabitEditScreen(edit.id)
+        },
+        content = {
+            Scaffold(
+                topBar = {
+                    TopAppBar(
+                        title = {
+                            Text(
+                                stringResource(selectedItem.title),
+                                maxLines = 1,
+                                overflow = TextOverflow.Ellipsis
+                            )
+                        },
+                        navigationIcon = {
+                            IconButton(onClick = { scope.launch { drawerState.open() } }) {
+                                Icon(
+                                    imageVector = Icons.Filled.Menu,
+                                    contentDescription = "Open Navigation Drawer"
+                                )
+                            }
+                        },
+                        actions = {
+                            if (selectedItem == Routes.Home) {
+                                IconButton(
+                                    onClick = {
+                                        showBottomSheet = true
+                                    }
+                                ) {
+                                    Icon(
+                                        imageVector = Icons.Filled.Search,
+                                        contentDescription = stringResource(R.string.search)
+                                    )
+                                }
+                            }
+                        }
+                    )
+                }
+            ) { paddingValues ->
+                NavHost(
+                    modifier = Modifier.padding(paddingValues),
+                    navController = navController,
+                    startDestination = Routes.Home
+                ) {
+                    composable<Routes.Home> {
+                        HabitsPagerScreen(
+                            onNavigate = { id ->
+                                navController.navigate(
+                                    Routes.HabitEdit(id)
+                                )
+                            },
+                            showBottomSheet = showBottomSheet,
+                            onShowBottomSheetChange = {
+                                showBottomSheet = it
+                            }
+                        )
+                    }
+                    composable<Routes.Info> {
+                        InfoScreen()
+                    }
+                    composable<Routes.HabitEdit> { backStackEntry ->
+                        val edit: Routes.HabitEdit = backStackEntry.toRoute()
+                        HabitEditScreen(
+                            habitId = edit.id,
+                            onBack = { navController.popBackStack() })
+                    }
+                }
             }
         }
-    }
+    )
 }
