+++ app/src/main/java/com/doubletapp_hw/db/AppDatabase.kt
@@ -4,9 +4,15 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import androidx.room.TypeConverter
+import androidx.room.TypeConverters
 import com.doubletapp_hw.Habit
+import com.squareup.wire.Instant
+import java.time.LocalDateTime
+import java.time.ZoneId
 
-@Database(entities = [Habit::class], version = 2, exportSchema = false)
+@Database(entities = [Habit::class], version = 1, exportSchema = false)
+@TypeConverters(LocalDateTimeConverter::class)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun habitDao(): HabitDao
 
@@ -17,12 +23,29 @@ abstract class AppDatabase : RoomDatabase() {
         fun getDatabase(context: Context): AppDatabase {
             return INSTANCE ?: synchronized(this) {
                 val instance = Room.databaseBuilder(
-                    context.applicationContext, AppDatabase::class.java, "habit_database"
-                ).fallbackToDestructiveMigration(true)
+                    context.applicationContext,
+                    AppDatabase::class.java,
+                    "habit_database"
+                )
                     .build()
                 INSTANCE = instance
                 instance
             }
         }
+    }
+}
+
+
+class LocalDateTimeConverter {
+    @TypeConverter
+    fun fromTimestamp(value: Long?): LocalDateTime? {
+        return value?.let {
+            Instant.ofEpochMilli(it).atZone(ZoneId.systemDefault()).toLocalDateTime()
+        }
+    }
+
+    @TypeConverter
+    fun dateToTimestamp(date: LocalDateTime?): Long? {
+        return date?.atZone(ZoneId.systemDefault())?.toInstant()?.toEpochMilli()
     }
 }
