--- app/src/main/java/com/doubletapp_hw/enums/HabitPriority.kt
@@ -1,9 +1,0 @@
-package com.doubletapp_hw.enums
-
-import com.doubletapp_hw.R
-
-enum class HabitPriority(val labelResId: Int) {
-    LOW(R.string.low),
-    MID(R.string.mid),
-    HIGH(R.string.hight);
-}
