--- app/src/main/java/com/doubletapp_hw/enums/HabitType.kt
@@ -1,8 +1,0 @@
-package com.doubletapp_hw.enums
-
-import com.doubletapp_hw.R
-
-enum class HabitType(val labelResId: Int) {
-    POSITIVE(R.string.positive),
-    NEGATIVE(R.string.negative);
-}
