--- app/src/main/java/com/doubletapp_hw/enums/SortingType.kt
@@ -1,9 +1,0 @@
-package com.doubletapp_hw.enums
-
-import com.doubletapp_hw.R
-
-enum class SortingType(val labelResId: Int) {
-    NAME(R.string.name),
-    DATE(R.string.date),
-    PRIORITY(R.string.priority)
-}
