--- app/src/main/java/com/doubletapp_hw/screens/NavDrawer.kt
@@ -1,132 +1,0 @@
-package com.doubletapp_hw.screens
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalDrawerSheet
-import androidx.compose.material3.ModalNavigationDrawer
-import androidx.compose.material3.NavigationDrawerItem
-import androidx.compose.material3.NavigationDrawerItemDefaults
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.rememberDrawerState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
-import coil.transform.CircleCropTransformation
-import com.doubletapp_hw.HabitApplication
-import com.doubletapp_hw.R
-import kotlinx.coroutines.launch
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun NavDrawer(
-    screen: Routes,
-    actions: @Composable RowScope.() -> Unit = {},
-    content: @Composable () -> Unit
-) {
-    val app = LocalContext.current.applicationContext as HabitApplication
-    val navController = app.localNavController.current
-    val scope = rememberCoroutineScope()
-    val drawerState = rememberDrawerState(DrawerValue.Closed)
-
-    val items = listOf(Routes.Home, Routes.Info)
-
-    ModalNavigationDrawer(
-        drawerState = drawerState,
-        drawerContent = {
-            ModalDrawerSheet {
-                Spacer(Modifier.height(16.dp))
-
-                Box(
-                    modifier = Modifier.fillMaxWidth(),
-                    contentAlignment = Alignment.TopCenter
-                ) {
-                    AsyncImage(
-                        model = ImageRequest.Builder(LocalContext.current)
-                            .data("https://images.techinsider.ru/upload/img_cache/1d1/1d1e22e5b38428a3a9875ea8677e5b9b_cropped_1332x1220.jpg")
-                            .crossfade(true)
-                            .transformations(CircleCropTransformation())
-                            .build(),
-                        contentDescription = "User Avatar",
-                        placeholder = painterResource(R.drawable.ic_launcher_foreground),
-                        error = painterResource(R.drawable.ic_launcher_foreground),
-                        modifier = Modifier
-                            .size(80.dp) // Размер аватара
-                            .clip(CircleShape) // Обводка по кругу
-                            .background(MaterialTheme.colorScheme.primary, CircleShape)
-                    )
-                }
-
-                Spacer(Modifier.height(24.dp))
-
-                items.forEach { item ->
-                    NavigationDrawerItem(
-                        icon = { Icon(item.icon, contentDescription = null) },
-                        label = { Text(stringResource(item.title)) },
-                        selected = item == screen,
-                        onClick = {
-                            scope.launch { drawerState.close() }
-                            navController.navigate(item)
-                        },
-                        modifier = Modifier.padding(NavigationDrawerItemDefaults.ItemPadding)
-                    )
-                }
-            }
-        },
-        content = {
-            Scaffold(
-                topBar = {
-                    TopAppBar(
-                        title = {
-                            Text(
-                                stringResource(screen.title),
-                                maxLines = 1,
-                                overflow = TextOverflow.Ellipsis
-                            )
-                        },
-                        navigationIcon = {
-                            IconButton(onClick = { scope.launch { drawerState.open() } }) {
-                                Icon(
-                                    imageVector = Icons.Filled.Menu,
-                                    contentDescription = "Open Navigation Drawer"
-                                )
-                            }
-                        },
-                        actions = actions
-                    )
-                }
-            ) {
-                Column(
-                    modifier = Modifier.padding(it),
-                ) {
-                    content()
-                }
-            }
-        }
-    )
-}
