+++ app/src/main/java/com/doubletapp_hw/viewModels/HabitEditViewModel.kt
@@ -8,23 +8,25 @@ import com.doubletapp_hw.Habit
 import com.doubletapp_hw.HabitRepository
 import kotlinx.coroutines.launch
 
-class HabitEditViewModel(
-    private val habitRepository: HabitRepository,
-    private val habitId: String
-) : ViewModel() {
+class HabitEditViewModel(private val habitRepository: HabitRepository) : ViewModel() {
     private val _habit = MutableLiveData<Habit>()
     val habit: LiveData<Habit> = _habit
 
-    init {
+    fun loadHabitById(id: String) {
         viewModelScope.launch {
-            _habit.value = habitRepository.getHabitById(habitId) ?: Habit()
+            _habit.value = habitRepository.getHabitById(id) ?: Habit()
         }
     }
 
     fun saveHabit(habit: Habit) {
         viewModelScope.launch {
-            habit.isSynced = false
             habitRepository.saveHabit(habit)
+        }
+    }
+
+    fun deleteHabit(habit: Habit) {
+        viewModelScope.launch {
+            habitRepository.deleteHabit(habit)
         }
     }
 }
