+++ build.gradle.kts
@@ -1,9 +1,9 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id("com.android.application") version "8.9.2" apply false
+    id("com.android.application") version "8.9.1" apply false
     id("org.jetbrains.kotlin.android") version "2.1.10" apply false
     id("com.google.devtools.ksp") version "2.1.10-1.0.31" apply false
     id("com.android.library") version "8.2.2" apply false
     id("org.jetbrains.kotlin.plugin.compose") version "2.1.10" apply false
-
+    //alias(libs.plugins.compose.compiler) apply false
 }
